clang shorten-64-to-32 warnings (#1170)
* alsa/omx/oss: fix shorten_64_to_32 warnings

* alsa: fix shorten_64_to_32 with long

+++ modules/alsa/alsa_play.c
@@ -54,7 +54,7 @@ static void auplay_destructor(void *arg)
 static void *write_thread(void *arg)
 {
 	struct auplay_st *st = arg;
-	int n;
+	long n;
 	int num_frames;
 
 	num_frames = st->prm.srate * st->prm.ptime / 1000;
